Update classification-QSAR-bioKom.knwf
In the "Summary" file, the names of the metrics were modified (metric + file name), the names of the feature selection methods in the "MethodFS" column were improved.
In the input parameters, the name "ClassName" was changed to "TargetFeature," the parameters "SpearmanCorrelation" and "Entropy" were added to specify the threshold used in the corresponding filter, and the parameter "Processors" was added to specify the number of chunks to be used in parallelization nodes.