Fixing t-SNE for the test sets
Fixing ERROR	 KNIME-Worker-37-Variable Loop End 3:2613:2631:2647:2640:2637 Node	 Active Scope End node in inactive branch not allowed.

It was added the use of try-catch disclosure to avoid that the workflow stops when building t-SNE graphics for the test sets. However, there is a bug in the KNIME platform related to the wrong working of the try-catch. See links below:

https://forum.knime.com/t/try-catch-active-scope-end-node-in-inactive-branch-not-allowed/82010

https://forum.knime.com/t/bug-active-scope-end-node-in-inactive-branch-not-allowed/3489